\name{networkDepth-package}
\alias{networkDepth-package}
\alias{networkDepth}
\docType{package}
\title{
\packageTitle{networkDepth}
}
\description{
\packageDescription{networkDepth}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{networkDepth}
\packageIndices{networkDepth}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{networkDepth}

Maintainer: \packageMaintainer{networkDepth}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
